chore: refactored info button/info-dialog and in the RUI
