+++ projects/ccf-rui/src/app/modules/right-sidebar/right-sidebar.module.ts
@@ -14,7 +14,7 @@ import { TagSearchModule } from '../../shared/components/tag-search/tag-search.m
 import { ReviewButtonModule } from '../review/review-button/review-button.module';
 import { RightSidebarComponent } from './right-sidebar.component';
 
-import { InfoButtonModule } from '../info/info-button/info-button.module';
+import { InfoButtonModule } from 'ccf-shared';
 
 
 @NgModule({
