+++ projects/ccf-rui/src/themes/common/_theming.scss
@@ -1,7 +1,7 @@
 @use '~@angular/material/theming' as *;
 @use '../../app/modules/header/header.component.theme' as header;
-@use '../../app/modules/info/info-dialog/info-dialog.component.theme.scss' as infoDialog;
-@use '../../app/modules/info/info-button/info-button.component.theme.scss' as infoButton;
+@use '../../../../ccf-shared/src/lib/components//info/info-dialog/info-dialog.component.theme.scss' as infoDialog;
+@use '../../../../ccf-shared/src/lib/components/info/info-button/info-button.component.theme.scss' as infoButton;
 @use '../../app/shared/components/drawer/drawer.theme' as drawer;
 @use '../../app/modules/right-sidebar/right-sidebar.component.theme.scss' as rightSidebar;
 @use '../../app/shared/components/block-size-input/block-size-input.component.theme.scss' as blockSizeInput;
