+++ projects/ccf-shared/src/lib/components/info/info-button/info-button.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { InfoButtonComponent } from './info-button.component';
-import { InfoDialogModule } from '../info-dialog/info-dialog.module';
 import { HttpClientModule } from '@angular/common/http';
+import { InfoDialogModule } from '../info-dialog/info-dialog.module';
 
 @NgModule({
   declarations: [InfoButtonComponent],
