+++ projects/ccf-shared/src/lib/components/info/info-dialog/info-dialog.component.scss
